Implemented  auth functionality

+++ db/db_note.py
@@ -1,14 +1,14 @@
 from sqlalchemy.orm.session import Session
-from schemas import NoteBase, NoteList
-from db.models import DbNote, DbNoteHistory
+from schemas import NoteBase
+from db.models import DbNote, DbNoteHistory, DbUser
 from fastapi import HTTPException
-from datetime import datetime
 
-def create_note(db:Session, request:NoteBase):
+
+def create_note(db:Session, request:NoteBase, current_user:DbUser):
     new_note = DbNote(
         title = request.title,
         description=request.description,
-        owner_id = request.owner_id
+        owner_id = current_user.id
     )
 
     db.add(new_note)
@@ -19,23 +19,31 @@ def create_note(db:Session, request:NoteBase):
 def get_all(db:Session):
     return db.query(DbNote).all()
 
-def get_one(db:Session, id:int):
+def get_one(db:Session, id:int, current_user:DbUser):
     note = db.query(DbNote).filter(DbNote.id==id).first()
     if not note:
         raise HTTPException(status_code=404, detail=f'Note {id} not found')
+    if note.owner != current_user:
+        raise HTTPException(status_code=403, detail=f'Forbidden')
     return note
 
-def get_note_history(db:Session, id:int):
+def get_note_history(db:Session, id:int, current_user:DbUser):
     note=db.query(DbNote).filter(DbNote.id==id).first()
     if not note:
         raise HTTPException(status_code=404, detail=f'Note {id} not found')
+    if note.owner != current_user:
+        raise HTTPException(status_code=403, detail=f'Forbidden')
+    if not note.history:
+        raise HTTPException(status_code=404, detail=f'History is empty')
     return note.history
 
-def update_note(db:Session, id:int, request:NoteBase):
+def update_note(db:Session, id:int, request:NoteBase, current_user:DbUser):
     note = db.query(DbNote).filter(DbNote.id==id)
-    
+
     if not note.first():
         raise HTTPException(status_code=404, detail=f'Note {id} not found')
+    if note.first().owner != current_user:
+        raise HTTPException(status_code=403, detail=f'Forbidden')
     
     note_history = DbNoteHistory(
         note_id = note.first().id,
@@ -57,10 +65,12 @@ def update_note(db:Session, id:int, request:NoteBase):
     return note.first()
 
 
-def delete_note(db:Session, id:int):
+def delete_note(db:Session, id:int, current_user:DbUser):
     note = db.query(DbNote).filter(DbNote.id==id).first()
     if not note:
         raise HTTPException(status_code=404, detail=f'Note {id} not found')
+    if note.owner != current_user:
+        raise HTTPException(status_code=403, detail=f'Forbidden')
     
     db.delete(note)
     db.commit()
